Add plug for role-based authorization
