+++ lib/plausible_web/plugs/authorize_stats_plug.ex
@@ -2,12 +2,11 @@ defmodule PlausibleWeb.AuthorizeStatsPlug do
   import Plug.Conn
   use Plausible.Repo
 
-  def init(options) do
-    options
-  end
+  def init([]), do: [:public, :viewer, :admin, :owner]
+  def init(allowed_roles), do: allowed_roles
 
-  def call(conn, _opts) do
-    site = Repo.get_by(Plausible.Site, domain: conn.params["domain"])
+  def call(conn, allowed_roles) do
+    site = Repo.get_by(Plausible.Site, domain: conn.params["domain"] || conn.params["website"])
     shared_link_auth = conn.params["auth"]
 
     shared_link_record =
@@ -17,17 +16,27 @@ defmodule PlausibleWeb.AuthorizeStatsPlug do
       PlausibleWeb.ControllerHelpers.render_error(conn, 404) |> halt
     else
       user_id = get_session(conn, :current_user_id)
-      role = user_id && Plausible.Sites.role(user_id, site)
+      membership_role = user_id && Plausible.Sites.role(user_id, site)
 
-      can_access =
-        site.public ||
-          (user_id && role) ||
-          (shared_link_auth && shared_link_record && shared_link_record.site_id == site.id)
+      role =
+        cond do
+          user_id && membership_role ->
+            membership_role
 
-      if !can_access do
-        PlausibleWeb.ControllerHelpers.render_error(conn, 404) |> halt
-      else
+          site.public ->
+            :public
+
+          shared_link_record && shared_link_record.site_id == site.id ->
+            :public
+
+          true ->
+            nil
+        end
+
+      if role in allowed_roles do
         merge_assigns(conn, site: site, current_user_role: role)
+      else
+        PlausibleWeb.ControllerHelpers.render_error(conn, 404) |> halt
       end
     end
   end
